codebase updates
stats methods error fix
hmmlearn helper function created
